Move DependencyFormatChecker to separate module
